+++ keras_spiking/__init__.py
@@ -4,6 +4,7 @@ __copyright__ = "2020-2020, Applied Brain Research"
 __license__ = "Free for non-commercial use; see LICENSE.rst"
 from keras_spiking.version import version as __version__
 
+from keras_spiking import callbacks, layers, regularizers
 from keras_spiking.layers import (
     Alpha,
     AlphaCell,